% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA_class.R
\name{MTA}
\alias{MTA}
\title{Class MTA}
\description{
Class MTA

Class MTA
}
\details{
Represent all metadata of a marker trait association
}
\examples{
 mta <- MTA$new(
   alleleID = "120515446-16-G/T",
   alleleSequence = "TGCAGAGGAAGAATCATGGGTTAGCTTGTTCAAATCTTCGGGTAACTAAACAAATAAAACGACTTGTGC",
   favorableAllele = "T"
 )

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alleleID}}{a string in the format of "chr-pos-allele1/allele2" for SNP MTA, and "chr-pos" for SIL MTA}

\item{\code{alleleSequence}}{a string of the allele sequence, with the same length as the reference genome sequence. The allele at the position of the locus tag must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{favorableAllele}}{a string of the favorable allele, which must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{mtaType}}{a string indicating the type of MTA, either "SNP" or "SIL"}

\item{\code{alleles}}{a vector of alleles for the MTA}

\item{\code{loc_tag_pos}}{a numeric indicating the position of the locus tag in the allele sequence for SNP MTA, and NULL for SIL MTA}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MTA-new}{\code{MTA$new()}}
\item \href{#method-MTA-find_homologs}{\code{MTA$find_homologs()}}
\item \href{#method-MTA-clone}{\code{MTA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-new"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MTA object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$new(alleleID, alleleSequence, favorableAllele)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alleleID}}{a string in the format of "chr-pos-allele1/allele2" for SNP MTA, and "chr-pos" for SIL MTA}

\item{\code{alleleSequence}}{a string of the allele sequence, with the same length as the reference genome sequence. The allele at the position of the locus tag must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{favorableAllele}}{a string of the favorable allele, which must be either allele 1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new MTA object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-find_homologs"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-find_homologs}{}}}
\subsection{Method \code{find_homologs()}}{
Find homologous loci in a \code{DArTSet} for the current MTA.

Uses Hamming distance between the MTA allele sequence and each sequence
in \code{dartset$db}. For SNP MTAs, candidate loci are further filtered to
keep only those matching the SNP position.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$find_homologs(dartset, min_distance = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dartset}}{A \code{DArTSet} object used as search space for homologs.}

\item{\code{min_distance}}{Maximum Hamming distance allowed for a sequence to be
considered homologous.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named numeric vector of Hamming distances for matching loci, or
\code{NA} when no homolog is found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
