% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTA_class.R
\name{MTA}
\alias{MTA}
\title{Class MTA}
\description{
Class MTA

Class MTA
}
\details{
Represent all metadata of a marker trait association
}
\examples{
 mta <- MTA$new(
   alleleID = "120515446-16-G/T",
   alleleSequence = "TGCAGAGGAAGAATCATGGGTTAGCTTGTTCAAATCTTCGGGTAACTAAACAAATAAAACGACTTGTGC",
   favorableAllele = "T"
 )


## ------------------------------------------------
## Method `MTA$print`
## ------------------------------------------------

print(MTA)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alleleID}}{a string in the format of "chr-pos-allele1/allele2" for SNP MTA, and "chr-pos" for SIL MTA}

\item{\code{alleleSequence}}{a string of the allele sequence, with the same length as the reference genome sequence. The allele at the position of the locus tag must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{favorableAllele}}{a string of the favorable allele, which must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{mtaType}}{a string indicating the type of MTA, either "SNP" or "SIL"}

\item{\code{alleles}}{a vector of alleles for the MTA}

\item{\code{loc_tag_pos}}{a numeric indicating the position of the locus tag in the allele sequence for SNP MTA, and NULL for SIL MTA}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MTA-new}{\code{MTA$new()}}
\item \href{#method-MTA-print}{\code{MTA$print()}}
\item \href{#method-MTA-clone}{\code{MTA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-new"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MTA object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$new(alleleID, alleleSequence, favorableAllele)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alleleID}}{a string in the format of "chr-pos-allele1/allele2" for SNP MTA, and "chr-pos" for SIL MTA}

\item{\code{alleleSequence}}{a string of the allele sequence, with the same length as the reference genome sequence. The allele at the position of the locus tag must be either allele1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}

\item{\code{favorableAllele}}{a string of the favorable allele, which must be either allele 1 or allele2 for SNP MTA, and must be 0 or 1 for SIL MTA.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new MTA object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-print"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a printout of the MTA object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{print(MTA)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MTA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MTA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MTA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
